+++ changelog.md
@@ -1,6 +1,9 @@
 # 1.1.5
-- Fixed "Mark function as Throws" when inside a lambda (The annotation got placed at the lambda rather than at the
-  function declaration)
+ - Fixed "Mark function as Throws" when inside a lambda (The annotation got placed at the lambda rather than at the function declaration)
+ - Added most of kotlin std lib functions as either callthough or ignore
+ - Fixed issues with java interopt
+ - Fixed issues with Throwable
+ - Fixed issues with indirect exceptions (via variables & method calls)
 
 # 1.1.4
 - disables throws annotator in test code / modules